fix plugins
