+++ app/src/main/java/com/lagradost/cloudstream3/ui/settings/extensions/PluginAdapter.kt
@@ -176,7 +176,8 @@ class PluginAdapter(
             if (data.isDownloaded) {
                 // On local plugins page the filepath is provided instead of url.
                 val plugin =
-                    PluginManager.urlPlugins[metadata.url] ?: PluginManager.plugins[metadata.url]
+                    (PluginManager.urlPlugins[metadata.url] ?: (PluginManager.plugins[metadata.url])) as? com.lagradost.cloudstream3.plugins.Plugin
+
                 if (plugin?.openSettings != null) {
                     binding.actionSettings.isVisible = true
                     binding.actionSettings.setOnClickListener {
